fix: ajuste em not found em controller e subvidindo helper/argon
